Add an `autoconvert` feature.
The feature is enabled by default and allows for operations between
quantities with different base units. `autoconvert` enabled is the same
functionality as prior versions of `uom`. Disabling the feature only
allows for quantities with the same base units to directly interact. The
feature has been added to account for current compiler limitations where
zero-cost code is not generated for non-floating point underlying
storage types. With the feature disabled more optimal code will be
generated for integral types. This feature will likely have no effect
and subsequently be deprecated and removed once const fn and
specialization are stabilized. Resolves #54.