Add separate testcrate
Allows testing without inheriting features from the uom build, is tested
when running `cargo test` automatically be being a workspace member, so
will be tested on travis and appveyor. Includes all the examples from
uom so that these examples are tested as if they were part of a separate
application.