Reduce number of storage types that are tested for performance reasons.
Compiler performance issues described in #52 are causing the `uom` build
to take a significant amount of time when all underlying storage types
are enabled. As additional quantities are added this performance gets
worse. This commit only tests a few types in order to keep the build
time within Appveyor and TravisCI limits.