Fix tests on `rustc` 1.20.0.
Use an internal `1.20.0` feature to disable `compile_fail` tests for
`rustc` 1.20.0. Additional `stable`, `beta`, and `nightly` features are
added for builds on the corresponding channels and remain unused. Note
that a channel's feature must be added to the `--features` list when
compiling manually as no channel detection is built into
`cargo`/`rustc`.

Tests are split into multiple `cargo test` runs because of limitations
with passing features to workspace crates.

Resolves #81.