Allow Clippy `excessive_precision` lint.
Conversion factors are specified based on the current best known factors
documented in NIST 811. Some of these factors exceed the precision of
`f64` and trigger the `excessive_precision` lint. For the time being the
lint is being disable and #40 will more accurately handle conversion
factors for all underlying storage types.