Correct `to_base` and `from_base` to be zero-cost for float storage types.
For a value, `v: Float`, adding `-0.0` is a no-op while adding `0.0`
will change the sign if `v` is `-0.0`. The opposite is true for
subtraction.

When no constant factor exists use an appropriated signed `0.0`. A new
enum, `ConstantOp`, is added to allow the `constant(op: ConstantOp)`
function to return an appropriately signed `0.0`.

   v
 0.0 + -0.0 =  0.0
-0.0 +  0.0 =  0.0 // v +  0.0 != v
-0.0 + -0.0 = -0.0
 0.0 - -0.0 =  0.0
-0.0 -  0.0 =  0.0
-0.0 - -0.0 =  0.0 // v - -0.0 != v