Ensure `UnwindSafe` and `RefUnwindSafe` are implemented.
Because both traits only exist in `std` new definitions are exported
when the `std` feature is not enabled. The `Dimension` and `Units`
require traits with these names to exist as there is no way to have
conditional trait inheritance based on feature selection.

This fix also simplifies ensuring that `Unpin` is implemented by using
trait inheritance instead of explicit impls.

Resolves #227.