Add explicit serde feature
This revision adds an explicit Cargo feature named serde. This feature is
analogous to the old use_serde feature, but uses features of Cargo's new
feature resolver to automatically enable serde features for num-* crates as
appropriate. The use_serde feature is now a (deprecated) alias for this new
serde feature.