Add libm dependency
This commit adds the `libm` optional feature (not enabled by default) to
turn the corresponding feature in the `num-traits` dependency crate. The
commit updates some default SI units and changes `Float` exporting
rules.

The `libm` crate enables math support for `#![no_std]` crates,
including trigonometric functions. Those are implemented for `Angle` and
`Ratio` quantities, and now are supported and usable in a wider area of
cases.