Improve handling of conversions.
Currently a conversion, which is a mandatory coefficient and an optional
constant, is matched with `$($conversion:expr),+`. This is overly
general because it can match more than two expressions.

This commit changes it to `$coefficient:expr $(, $constant:expr)?`,
which is exact and clearer. It also removes the need for the
`@coefficient` rule because `$coefficient` can be used directly, and the
`@constant` rule becomes simpler.

Finally, it's faster to compile. A `cargo check` build drops from 12.2s
to 11.8s on my machine.