Fix conversion of complex types, add From
Move the PartialOrd trait from the ConversionFactor and onto T. T is
usually f64 or something compatible (at the very least a scalar),
while V might be Complex. We really only need to compare the scaling
factor by the scalar (of type T). This didn't matter much when both V
and T where scalar, but it broke down for Complex.

This means Conversion<V> no longer has a default implementation for
conversion(), which must return something of the underlying type V.

It's quite common to start with some values that through
transformations end up complex, such as in signal processing. To
support this I added into_complex() (which is very clear in its
intention) and From<Quantity<V>> -> Quantity<Complex<V>>. This
supports code like:

    // m = meter, s = second
    let c1 = (m / s).into_complex();
    let c2 = uom::complex64::Velcity::from(m /s);
    assert!(c1 == c2);

into_complex() on an already.complex value is a no-op, to make it
easier to write generic code, e.g.:

    u.into_complex() == v.into_complex().into_complex()

The rest of the patch is dealing with the outcome; propagating trait
constraints and types. Notably, functions that call
{to,from,change}_base must provide +-*/ in order to support the
conversion.

There's still some work left, but complex types are at least usable
now, and more tests pass. What's missing:

1. The test suite needs some polish to better support
   complex, particularly the quickcheck tests. Scaling and
   transforming between bases when using very large values quickly
   overflow into inf or NaN, and the filters for this behaviour for
   non-complex numbers are not sufficient.
2. To/from string is not implemented. We probably don't want to use
   the default formatter for Complex, but rather output and parse
   a (a+bi) or (a+bj) format. This is not implemented and so these
   tests fail.
3. Some tests in si/ break because complex scaling does more floating
   point operations and introduce more imprecision. It might be
   sufficient to change some of these from eq to approx_eq.
4. The the rem/% operator which should probably be removed or at least
   carefully property checked for complex storage.